+++ packages/sv-utils/src/tests/js/exports/named-export-with-existing-function/output.ts
@@ -1,0 +1,2 @@
+export function existing() {}
+export const variable = {};